+++ src/markdown.tsx
@@ -1,11 +1,9 @@
-import { BaseWidget } from '@tinystacks/ops-core';
-import { Widget } from '@tinystacks/ops-model';
 import React from 'react';
+import { BaseWidget } from '@tinystacks/ops-core';
 import { default as ReactMarkdown } from 'react-markdown';
 import { default as ChakraUIRenderer } from 'chakra-ui-markdown-renderer';
 import { Box } from '@chakra-ui/react';
-
-export type MarkdownProps = Widget & { markdown: string }
+import { Markdown as MarkdownProps } from './ops-types.js';
 
 export class Markdown extends BaseWidget {
   markdown: string;