+++ test/utils/find-provider.test.ts
@@ -1,6 +1,5 @@
 import { findProvider } from '../../src/utils/find-provider.js';
 import { GithubCredentialsProvider } from '../../src/providers/github-credentials-provider.js';
-import { BaseProvider } from '@tinystacks/ops-core';
 import { MockProvider } from '../mocks/MockProvider.js';
 
 describe('findProvider', () => {